MVP/backend/pyproject.toml
[project]
name = "world-of-shadows-backend"
version = "1.0.0"
description = "Flask backend for World of Shadows game platform"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.6,<4",
    "flask-cors>=4.0.0,<5",
    "flask-sqlalchemy>=3.1.0,<4",
    "flask-jwt-extended>=4.6.0,<5",
    "flask-limiter>=3.5.0,<4",
    "flask-migrate>=4.0.0,<5",
    "flask-wtf>=1.2.1,<2",
    "flask-mail>=0.10.0,<1",
    "werkzeug>=3.1.5,<4",
    "alembic>=1.18.0,<2",
    "python-dotenv>=1.0.0,<2",
    "markdown>=3.5.0,<4",
    "bleach>=6.0.0,<7",
    "gunicorn>=21.2.0,<23",
    "httpx>=0.27.0,<1",
    "email-validator>=2.0.0,<3",
    "pydantic>=2.0.0,<3",
    "fastapi>=0.115.0,<1",
    "langchain>=1.2.0,<2",
    "langgraph>=1.0.3,<2",
    "numpy>=1.24.0,<3",
    "fastembed>=0.3.0,<1",
    "requests>=2.31.0,<3",
    "pyyaml>=6.0,<7",
    "marshmallow>=3.20.0,<4",
    "cryptography>=45.0.7,<47",
    "pyOpenSSL>=25.3.0,<27",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0,<9",
    "pytest-asyncio>=0.21,<1",
    "pytest-cov>=4.0,<6",
    "pytest-timeout>=2.1",
    "anyio>=4.0,<5",
    "exceptiongroup>=1.2.0; python_version < '3.11'",
    "uvicorn>=0.30,<1",
]
dev = [
    "pytest>=7.0,<9",
    "pytest-asyncio>=0.21,<1",
    "pytest-cov>=4.0,<6",
    "pytest-timeout>=2.1",
    "anyio>=4.0,<5",
    "uvicorn>=0.30,<1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "strict"
addopts = "-v --tb=short --cov=app --cov-report=term-missing --cov-config=.coveragerc --cov-branch"
filterwarnings = [
    "ignore::sqlalchemy.exc.LegacyAPIWarning",
]

[tool.coverage.run]
source = ["app"]
branch = true
omit = [
    "*/site-packages/*",
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
    "@overload",
]
precision = 2
skip_covered = false
skip_empty = true

[tool.coverage.paths]
source = [
    "app",
    "*/app",
    "/mnt/c/*/app",
    "C:\\Users\\*\\*\\app",
]
